refactor: mock data, api 파일 분리
