+++ src/mocks/api/detailHandler.ts
@@ -1,45 +1,13 @@
 import { BASE_URL } from '@/api';
 import { PathParams, rest, RestRequest } from 'msw';
 import dayjs from 'dayjs';
+import { DETAILS } from '../data/detailData';
 
-/**
- * 상세 내역 생성시 데이터
- * groupId: number
- * "nickname" : String-"${팀원이름}"
- * "date" : String-"${사유 발생 날짜}" 23.06.01
- * "amount" : int-"${금액}" 1000
- * "ground" : String-"${사유}" 지각 | 결석 | 과제 안 함 | 기타
- * "memo" : String-"${메모}" 밥먹다 지각
- * "situation" : String-"${납부 여부}" 미납, 완납, 확인중
- */
+let details = DETAILS;
 
 const getSearchParams = (req: RestRequest<never, PathParams<string>>, param: string) => {
   return req.url.searchParams.get(param);
 };
-
-type Ground = '지각' | '결석' | '과제 안 함' | '기타';
-type Situation = '미납' | '완납' | '확인중';
-
-type EventInfo = {
-  eventId: number;
-  groupId: number;
-  nickname: string;
-  date: string;
-  amount: number;
-  ground: Ground;
-  memo: string;
-  situation: Situation;
-};
-
-let details: EventInfo[] = [
-  { eventId: 1, groupId: 1, nickname: '종현', date: '2023.06.25', amount: 10_000, ground: '지각', memo: '', situation: '미납' },
-  { eventId: 2, groupId: 1, nickname: '정민', date: '2023.06.30', amount: 10_000, ground: '결석', memo: '', situation: '완납' },
-  { eventId: 3, groupId: 1, nickname: '윤하나', date: '2023.07.01', amount: 10_000, ground: '지각', memo: '', situation: '미납' },
-  { eventId: 4, groupId: 1, nickname: '가람', date: '2023.06.25', amount: 10_000, ground: '과제 안 함', memo: '', situation: '확인중' },
-  { eventId: 5, groupId: 1, nickname: '나경', date: '2023.07.04', amount: 10_000, ground: '지각', memo: '', situation: '확인중' },
-  { eventId: 6, groupId: 1, nickname: '현교', date: '2023.07.04', amount: 10_000, ground: '지각', memo: '', situation: '미납' },
-  { eventId: 7, groupId: 1, nickname: '재민', date: '2023.07.04', amount: 10_000, ground: '지각', memo: '', situation: '완납' },
-];
 
 const createDetail = () => {
   return rest.post(BASE_URL + '/api/event/penalty', async (req, res, ctx) => {
